+++ lib/common/exceptions/api_excetion.dart
@@ -1,0 +1,5 @@
+class ApiExcetion implements Exception {
+  final String message;
+
+  ApiExcetion({required this.message});
+}
