+++ lib/common/failure/failure.dart
@@ -1,0 +1,5 @@
+abstract class Failure {
+  final String message;
+
+  Failure({required this.message});
+}
