+++ lib/models/failures/api_failure.dart
@@ -1,0 +1,8 @@
+import 'package:bookmywarehouse/common/failure/failure.dart';
+
+class ApiFailure implements Failure {
+  @override
+  final String message;
+
+  ApiFailure({required this.message});
+}
