+++ lib/src/auth/widgets/sign_in_button.dart
