+++ lib/src/auth/widgets/singin_with_phn.dart
@@ -1,5 +1,5 @@
 import 'package:bookmywarehouse/constants/color/base_color.dart';
-import 'package:bookmywarehouse/ui_components/login_signup_page/otp_screen.dart';
+import 'package:bookmywarehouse/src/auth/widgets/otp_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
