+++ lib/src/auth/widgets/welcome.dart
