+++ lib/ui_components/fotgot_password/email_confirmation.dart
@@ -1,5 +1,5 @@
 import 'package:bookmywarehouse/constants/color/base_color.dart';
-import 'package:bookmywarehouse/widgets/login_screen.dart';
+import 'package:bookmywarehouse/src/auth/pages/login_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:lottie/lottie.dart';
