+++ lib/ui_components/fotgot_password/setpassword.dart
@@ -1,6 +1,6 @@
 import 'package:bookmywarehouse/constants/color/base_color.dart';
 import 'package:bookmywarehouse/ui_components/fotgot_password/recover_text.dart';
-import 'package:bookmywarehouse/widgets/login_screen.dart';
+import 'package:bookmywarehouse/src/auth/pages/login_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
