+++ lib/widgets/navigation_page.dart
@@ -1,7 +1,7 @@
 import 'package:bookmywarehouse/constants/color/base_color.dart';
 import 'package:bookmywarehouse/widgets/chat_screen.dart';
 import 'package:bookmywarehouse/widgets/explore_screen.dart';
-import 'package:bookmywarehouse/widgets/home_screen.dart';
+import 'package:bookmywarehouse/src/home/pages/home_screen.dart';
 import 'package:bookmywarehouse/widgets/profilescreen.dart';
 import 'package:bookmywarehouse/widgets/savded_screen.dart';
 import 'package:flutter/material.dart';
